Moved Heater code to ktc_heater.py
